More refactoring, add back jar and run scripts for ease of use
