+++ RunWindows.cmd
@@ -1,0 +1,1 @@
+java -jar Landscape_Editor.jar
