+++ build.gradle
@@ -7,7 +7,18 @@ group "org.openrsc"
 version "1.0-SNAPSHOT"
 
 application {
-    mainClassName = "org.openrsc.editor.GUI"
+    mainClassName = "org.openrsc.editor.Main"
+}
+
+jar {
+    manifest {
+        attributes(
+                'Main-Class': 'org.openrsc.editor.Main'
+        )
+    }
+    from {
+        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
+    }
 }
 
 repositories {
@@ -16,7 +27,19 @@ repositories {
 }
 
 dependencies {
-    implementation "xpp3:xpp3:1.1.4c"
-    implementation "com.thoughtworks.xstream:xstream:1.4.9"
-
+    compile "xpp3:xpp3:1.1.4c"
+    compile "com.thoughtworks.xstream:xstream:1.4.9"
 }
+
+task copyJar(type: Copy) {
+    from jar // here it automatically reads jar file produced from jar task
+    into './'
+    rename(new Transformer<String, String>() {
+        @Override
+        String transform(String s) {
+            return name.substring(0, name.lastIndexOf("-")) + ".jar"
+        }
+    })
+}
+
+tasks.build.finalizedBy copyJar
