+++ src/main/java/org/openrsc/editor/Constants.java
@@ -1,0 +1,5 @@
+package org.openrsc.editor;
+
+public class Constants {
+    public static final long SLEEP_DELAY_MS = 8;
+}
