+++ src/main/java/org/openrsc/editor/Main.java
@@ -1,0 +1,11 @@
+package org.openrsc.editor;
+
+import org.openrsc.editor.gui.MainWindow;
+
+public class Main {
+    public static MainWindow mainWindow;
+
+    public static void main(String[] args) {
+        mainWindow = new MainWindow();
+    }
+}
