+++ src/main/java/org/openrsc/editor/MoveDirection.java
@@ -1,0 +1,8 @@
+package org.openrsc.editor;
+
+public enum MoveDirection {
+    UP,
+    DOWN,
+    LEFT,
+    RIGHT
+}
