+++ src/main/resources/xml/GameObjectLoc.xml.gz
