+++ src/main/resources/xml/ItemLoc.xml.gz
