+++ src/main/resources/xml/NpcLoc.xml.gz
