+++ src/main/resources/xml/item.txt
