+++ src/main/resources/xml/npc.txt
