+++ src/main/resources/xml/objects.txt
